+++ ElvUI_EltreumUI/Core/Options.lua
@@ -102,6 +102,7 @@ local DONATORS = {
 	'|cffB50909Rocket Surgery|r',
 	'|cffB50909Embee|r',
 	'|cffB50909mercenariosgx|r',
+	'|cffB50909Deathclaw|r',
 }
 
 local TRANSLATORS = {
@@ -751,7 +752,7 @@ function ElvUI_EltreumUI:Configtable()
 	ElvUI_EltreumUI.Options.args.cvars.args.othercvars.args.camera = ACH:Group(L["Camera"], nil, 2, "tab")
 	ElvUI_EltreumUI.Options.args.cvars.args.othercvars.args.camera.args.description1 = ACH:Description(" ", 1, nil, 'Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\EltreumHeader', nil, 3240, 1, "full")
 	ElvUI_EltreumUI.Options.args.cvars.args.othercvars.args.camera.args.cameraFOV = ACH:Range(L["Camera Field of View"], L["This allows you to zoom out further with the camera to increase the field of view."], 2, { min = 50, max = 90, step = 1 }, 'full', function() return E.db.ElvUI_EltreumUI.cvars.cameraFOV end, function(_, value) E.db.ElvUI_EltreumUI.cvars.cameraFOV = value SetCVar('camerafov', value) end, nil)
-	ElvUI_EltreumUI.Options.args.cvars.args.othercvars.args.camera.args.cameraDistanceMaxZoomFactor = ACH:Range(L["Camera Distance Max Zoom Factor"], L["Maximum Camera Zoom Out"], 2, { min = 1, max = function() if not (E.Cata or E.Classic) then return 2.6 else return 3.4 end end, step = 0.1 }, 'full', function() return E.db.ElvUI_EltreumUI.cvars.cameraDistanceMaxZoomFactor end, function(_, value) E.db.ElvUI_EltreumUI.cvars.cameraDistanceMaxZoomFactor = value SetCVar('cameraDistanceMaxZoomFactor', value) end)
+	ElvUI_EltreumUI.Options.args.cvars.args.othercvars.args.camera.args.cameraDistanceMaxZoomFactor = ACH:Range(L["Camera Distance Max Zoom Factor"], L["Maximum Camera Zoom Out"], 2, { min = 1, max = function() if not (E.Cata or E.Classic) then return 2.6 else return 4 end end, step = 0.1 }, 'full', function() return E.db.ElvUI_EltreumUI.cvars.cameraDistanceMaxZoomFactor end, function(_, value) E.db.ElvUI_EltreumUI.cvars.cameraDistanceMaxZoomFactor = value SetCVar('cameraDistanceMaxZoomFactor', value) end)
 	ElvUI_EltreumUI.Options.args.cvars.args.othercvars.args.nameplates = ACH:Group(L["Nameplates"], nil, 2, "tab")
 	ElvUI_EltreumUI.Options.args.cvars.args.othercvars.args.nameplates.args.description1 = ACH:Description(" ", 1, nil, 'Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\EltreumHeader', nil, 3240, 1, "full")
 	ElvUI_EltreumUI.Options.args.cvars.args.othercvars.args.nameplates.args.SoftTargetInteract = ACH:Toggle(L["Soft Target Interact"], L["Enable Soft Target Interactions"], 2, nil, false,'full',
@@ -1741,6 +1742,8 @@ function ElvUI_EltreumUI:Configtable()
 	ElvUI_EltreumUI.Options.args.unitframes.args.general.args.appearance.args.skinblizzraid = ACH:Toggle(E.NewSign..L["Skin Blizzard Raid Frames"], L["Adds Gradient, Custom Textures, Shadows, Font and Role Icons to Blizzard Raid Frames"], 12, nil, false,'full', function() return E.db.ElvUI_EltreumUI.unitframes.blizzardraidframes end, function(_, value) E.db.ElvUI_EltreumUI.unitframes.blizzardraidframes = value E:StaticPopup_Show('CONFIG_RL') end, function() return not E.db.ElvUI_EltreumUI.unitframes.UFmodifications end)
 	ElvUI_EltreumUI.Options.args.unitframes.args.general.args.appearance.args.description6 = ACH:Description(L["Combat Indicator"], 13, nil, 'Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\EltreumHeader', nil, 3240, 1, "full")
 	ElvUI_EltreumUI.Options.args.unitframes.args.general.args.appearance.args.skincombaticons = ACH:Toggle(E.NewSign..L["Change Combat Indicators to be class based"], nil, 14, nil, false,'full', function() return E.db.ElvUI_EltreumUI.unitframes.classcombaticons end, function(_, value) E.db.ElvUI_EltreumUI.unitframes.classcombaticons = value end, function() return not E.db.ElvUI_EltreumUI.unitframes.UFmodifications end)
+	ElvUI_EltreumUI.Options.args.unitframes.args.general.args.appearance.args.description7 = ACH:Description(L["Heal Prediction"], 15, nil, 'Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\EltreumHeader', nil, 3240, 1, "full")
+	ElvUI_EltreumUI.Options.args.unitframes.args.general.args.appearance.args.healpredictionskin = ACH:Toggle(E.NewSign..L["Use ElvUI Unitframe Texture"], nil, 16, nil, false,'full', function() return E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.enableHealComm end, function(_, value) E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.enableHealComm = value end, function() return not E.db.ElvUI_EltreumUI.unitframes.UFmodifications end)
 	ElvUI_EltreumUI.Options.args.unitframes.args.general.args.backdrop = ACH:Group(L["Health"], nil, 2, "tab")
 	ElvUI_EltreumUI.Options.args.unitframes.args.general.args.backdrop.args.description1 = ACH:Description(" ", 1, nil, 'Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\EltreumHeader', nil, 3240, 1, "full")
 	ElvUI_EltreumUI.Options.args.unitframes.args.general.args.backdrop.args.hide = ACH:Toggle(L["Hide Backdrop"], L["Sets the Backdrop as Transparent/Hidden"], 2, nil, false,'full', function() return E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.backdrophidden end, function(_, value) E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.backdrophidden = value E:StaticPopup_Show('CONFIG_RL') end)
@@ -1976,6 +1979,8 @@ function ElvUI_EltreumUI:Configtable()
 		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.targettargettargettexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.playertexture
 		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.bosstexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.playertexture
 		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.focustexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.playertexture
+		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.raidtexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.playertexture
+		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.partytexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.playertexture
 		ElvUI_EltreumUI:GradientColorTableUpdate()
 	end, nil, true)
 	ElvUI_EltreumUI.Options.args.unitframes.args.texture.args.unitframes.args.gap1 = ACH:Description('', 3, nil)
@@ -1988,6 +1993,8 @@ function ElvUI_EltreumUI:Configtable()
 		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.targettargettargettexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.castbartexture
 		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.bosstexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.castbartexture
 		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.focustexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.castbartexture
+		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.raidtexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.castbartexture
+		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.partytexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.castbartexture
 		ElvUI_EltreumUI:GradientColorTableUpdate()
 	end, nil, true)
 	ElvUI_EltreumUI.Options.args.unitframes.args.texture.args.unitframes.args.gap2 = ACH:Description('', 6, nil)
@@ -2000,6 +2007,8 @@ function ElvUI_EltreumUI:Configtable()
 		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.targettargettargettexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.targettexture
 		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.bosstexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.targettexture
 		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.focustexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.targettexture
+		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.raidtexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.targettexture
+		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.partytexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.targettexture
 		ElvUI_EltreumUI:GradientColorTableUpdate()
 	end, nil, true)
 	ElvUI_EltreumUI.Options.args.unitframes.args.texture.args.unitframes.args.gap3 = ACH:Description('', 9, nil)
@@ -2012,6 +2021,8 @@ function ElvUI_EltreumUI:Configtable()
 		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.targettargettargettexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.targettargettexture
 		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.bosstexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.targettargettexture
 		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.focustexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.targettargettexture
+		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.raidtexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.targettargettexture
+		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.partytexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.targettargettexture
 		ElvUI_EltreumUI:GradientColorTableUpdate()
 	end, nil, true)
 	ElvUI_EltreumUI.Options.args.unitframes.args.texture.args.unitframes.args.gap4 = ACH:Description('', 12, nil)
@@ -2024,6 +2035,8 @@ function ElvUI_EltreumUI:Configtable()
 		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.pettexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.targettargettargettexture
 		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.bosstexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.targettargettargettexture
 		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.focustexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.targettargettargettexture
+		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.raidtexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.targettargettargettexture
+		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.partytexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.targettargettargettexture
 		ElvUI_EltreumUI:GradientColorTableUpdate()
 	end, nil, true)
 	ElvUI_EltreumUI.Options.args.unitframes.args.texture.args.unitframes.args.gap5 = ACH:Description('', 15, nil)
@@ -2036,6 +2049,8 @@ function ElvUI_EltreumUI:Configtable()
 		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.targettargettexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.focustexture
 		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.pettexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.focustexture
 		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.bosstexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.focustexture
+		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.raidtexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.focustexture
+		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.partytexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.focustexture
 		ElvUI_EltreumUI:GradientColorTableUpdate()
 	end, nil, true)
 	ElvUI_EltreumUI.Options.args.unitframes.args.texture.args.unitframes.args.gap6 = ACH:Description('', 18, nil)
@@ -2048,6 +2063,8 @@ function ElvUI_EltreumUI:Configtable()
 		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.targettargettexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.pettexture
 		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.focustexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.pettexture
 		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.bosstexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.pettexture
+		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.raidtexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.pettexture
+		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.partytexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.pettexture
 		ElvUI_EltreumUI:GradientColorTableUpdate()
 	end, nil, true)
 	ElvUI_EltreumUI.Options.args.unitframes.args.texture.args.unitframes.args.gap7 = ACH:Description('', 21, nil)
@@ -2060,9 +2077,38 @@ function ElvUI_EltreumUI:Configtable()
 		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.targettargettexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.bosstexture
 		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.focustexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.bosstexture
 		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.pettexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.bosstexture
+		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.raidtexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.bosstexture
+		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.partytexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.bosstexture
 		ElvUI_EltreumUI:GradientColorTableUpdate()
 	end, nil, true, nil, nil, nil, nil, E.Classic)
 	ElvUI_EltreumUI.Options.args.unitframes.args.texture.args.unitframes.args.gap8 = ACH:Description('', 24, nil)
+	ElvUI_EltreumUI.Options.args.unitframes.args.texture.args.unitframes.args.party = ACH:SharedMediaStatusbar(L["Party"], L["Select a Texture"], 25, "double", function() return E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.partytexture end, function(_,key) E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.partytexture = key ElvUI_EltreumUI:GradientColorTableUpdate() end, function() return not E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.enable end)
+	ElvUI_EltreumUI.Options.args.unitframes.args.texture.args.unitframes.args.applytoall9 = ACH:Execute(L["Apply To All"], nil, 26, function()
+		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.playertexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.partytexture
+		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.targettargettargettexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.partytexture
+		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.castbartexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.partytexture
+		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.targettexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.partytexture
+		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.targettargettexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.partytexture
+		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.focustexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.partytexture
+		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.bosstexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.partytexture
+		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.pettexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.partytexture
+		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.raidtexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.partytexture
+		ElvUI_EltreumUI:GradientColorTableUpdate()
+	end, nil, true)
+	ElvUI_EltreumUI.Options.args.unitframes.args.texture.args.unitframes.args.gap9 = ACH:Description('', 27, nil)
+	ElvUI_EltreumUI.Options.args.unitframes.args.texture.args.unitframes.args.raid = ACH:SharedMediaStatusbar(L["Raid"], L["Select a Texture"], 28, "double", function() return E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.raidtexture end, function(_,key) E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.raidtexture = key ElvUI_EltreumUI:GradientColorTableUpdate() end, function() return not E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.enable end)
+	ElvUI_EltreumUI.Options.args.unitframes.args.texture.args.unitframes.args.applytoall11 = ACH:Execute(L["Apply To All"], nil, 29, function()
+		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.playertexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.raidtexture
+		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.targettargettargettexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.raidtexture
+		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.castbartexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.raidtexture
+		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.targettexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.raidtexture
+		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.targettargettexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.raidtexture
+		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.focustexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.raidtexture
+		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.bosstexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.raidtexture
+		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.pettexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.raidtexture
+		E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.partytexture = E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.raidtexture
+		ElvUI_EltreumUI:GradientColorTableUpdate()
+	end, nil, true)
 	ElvUI_EltreumUI.Options.args.unitframes.args.texture.args.classes = ACH:Group(L["Classes"], nil, 3, "tab")
 	ElvUI_EltreumUI.Options.args.unitframes.args.texture.args.classes.args.description1 = ACH:Description(L["Class Textures"], 1, nil, 'Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\EltreumHeader', nil, 3240, 1, "full")
 	ElvUI_EltreumUI.Options.args.unitframes.args.texture.args.classes.args.dk = ACH:SharedMediaStatusbar(L["Death Knight"], L["Select a Texture"], 2, "double", function() return E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.deathknighttexture end, function(_,key) E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.deathknighttexture = key ElvUI_EltreumUI:GradientColorTableUpdate() end, function() return (not E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.enable) or E.db.ElvUI_EltreumUI.unitframes.ufcustomtexture.noclasstexture end, E.Classic)
@@ -2961,6 +3007,9 @@ function ElvUI_EltreumUI:Configtable()
 		["TYPE20"] = E:TextureString('Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\Nameplates\\reaper',':20:20'),
 		["TYPE21"] = E:TextureString('Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\Nameplates\\reaper2',':20:20'),
 		["TYPE22"] = E:TextureString('Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\Nameplates\\skull10',':20:20'),
+		["TYPE23"] = E:TextureString('Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\Nameplates\\skull11',':20:20'),
+		["TYPE24"] = E:TextureString('Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\Nameplates\\skull12',':20:20'),
+		["TYPE25"] = E:TextureString('Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\Nameplates\\skull13',':20:20'),
 		["CUSTOM"] = 'Custom',
 	}, false, nil, function() return E.db.ElvUI_EltreumUI.nameplates.classification.icontypeboss end, function(_,value) E.db.ElvUI_EltreumUI.nameplates.classification.icontypeboss = value end, function() return not E.db.ElvUI_EltreumUI.nameplates.classification.enable end)
 	ElvUI_EltreumUI.Options.args.nameplates.args.classification.args.boss.args.select.style = "radio"
@@ -2990,6 +3039,9 @@ function ElvUI_EltreumUI:Configtable()
 		["TYPE20"] = E:TextureString('Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\Nameplates\\reaper',':20:20'),
 		["TYPE21"] = E:TextureString('Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\Nameplates\\reaper2',':20:20'),
 		["TYPE22"] = E:TextureString('Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\Nameplates\\skull10',':20:20'),
+		["TYPE23"] = E:TextureString('Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\Nameplates\\skull11',':20:20'),
+		["TYPE24"] = E:TextureString('Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\Nameplates\\skull12',':20:20'),
+		["TYPE25"] = E:TextureString('Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\Nameplates\\skull13',':20:20'),
 		["CUSTOM"] = 'Custom',
 	}, false, nil, function() return E.db.ElvUI_EltreumUI.nameplates.classification.icontypeelite end, function(_,value) E.db.ElvUI_EltreumUI.nameplates.classification.icontypeelite = value end, function() return not E.db.ElvUI_EltreumUI.nameplates.classification.enable end)
 	ElvUI_EltreumUI.Options.args.nameplates.args.classification.args.elite.args.select.style = "radio"
@@ -3019,6 +3071,9 @@ function ElvUI_EltreumUI:Configtable()
 		["TYPE20"] = E:TextureString('Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\Nameplates\\reaper',':20:20'),
 		["TYPE21"] = E:TextureString('Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\Nameplates\\reaper2',':20:20'),
 		["TYPE22"] = E:TextureString('Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\Nameplates\\skull10',':20:20'),
+		["TYPE23"] = E:TextureString('Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\Nameplates\\skull11',':20:20'),
+		["TYPE24"] = E:TextureString('Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\Nameplates\\skull12',':20:20'),
+		["TYPE25"] = E:TextureString('Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\Nameplates\\skull13',':20:20'),
 		["CUSTOM"] = 'Custom',
 	}, false, nil, function() return E.db.ElvUI_EltreumUI.nameplates.classification.icontyperare end, function(_,value) E.db.ElvUI_EltreumUI.nameplates.classification.icontyperare = value end, function() return not E.db.ElvUI_EltreumUI.nameplates.classification.enable end)
 	ElvUI_EltreumUI.Options.args.nameplates.args.classification.args.rare.args.select.style = "radio"
@@ -3048,6 +3103,9 @@ function ElvUI_EltreumUI:Configtable()
 		["TYPE20"] = E:TextureString('Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\Nameplates\\reaper',':20:20'),
 		["TYPE21"] = E:TextureString('Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\Nameplates\\reaper2',':20:20'),
 		["TYPE22"] = E:TextureString('Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\Nameplates\\skull10',':20:20'),
+		["TYPE23"] = E:TextureString('Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\Nameplates\\skull11',':20:20'),
+		["TYPE24"] = E:TextureString('Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\Nameplates\\skull12',':20:20'),
+		["TYPE25"] = E:TextureString('Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\Nameplates\\skull13',':20:20'),
 		["CUSTOM"] = 'Custom',
 	}, false, nil, function() return E.db.ElvUI_EltreumUI.nameplates.classification.icontyperareelite end, function(_,value) E.db.ElvUI_EltreumUI.nameplates.classification.icontyperareelite = value end, function() return not E.db.ElvUI_EltreumUI.nameplates.classification.enable end)
 	ElvUI_EltreumUI.Options.args.nameplates.args.classification.args.rareelite.args.select.style = "radio"
@@ -3241,7 +3299,7 @@ function ElvUI_EltreumUI:Configtable()
 		[2] = L["Version 2"],
 	}, false, nil, function() return E.db.ElvUI_EltreumUI.skins.armorycrestversion end, function(_, value) E.db.ElvUI_EltreumUI.skins.armorycrestversion = value E:StaticPopup_Show('CONFIG_RL') end, function() return not E.db.ElvUI_EltreumUI.skins.expandarmorycrest end)
 	ElvUI_EltreumUI.Options.args.skins.args.character.args.panel.args.armory.args.addcrestversion.style = "radio"
-	ElvUI_EltreumUI.Options.args.skins.args.character.args.panel.args.armory.args.crestalpha = ACH:Range(L["Art Alpha"], L["Change the transparency of the Art"], 5, { min = 0.01, max = 1, step = 0.01 }, "full", function() return E.db.ElvUI_EltreumUI.skins.expandarmorybgalpha end, function(_, value) E.db.ElvUI_EltreumUI.skins.expandarmorybgalpha = value if _G["EltruismCharacterBG"] then _G["EltruismCharacterBG"]:SetAlpha(value) end end, function() return not E.db.ElvUI_EltreumUI.skins.expandarmorybg or not E.db.ElvUI_EltreumUI.skins.expandarmorycrest end)
+	ElvUI_EltreumUI.Options.args.skins.args.character.args.panel.args.armory.args.bgalpha = ACH:Range(L["Art Alpha"], L["Change the transparency of the Art"], 5, { min = 0.01, max = 1, step = 0.01 }, "full", function() return E.db.ElvUI_EltreumUI.skins.expandarmorybgalpha end, function(_, value) E.db.ElvUI_EltreumUI.skins.expandarmorybgalpha = value if _G["EltruismCharacterBG"] then _G["EltruismCharacterBG"]:SetAlpha(value) end end, function() return not E.db.ElvUI_EltreumUI.skins.expandarmorybg end)
 	ElvUI_EltreumUI.Options.args.skins.args.character.args.panel.args.armory.args.armorybgtextureselect = ACH:Select(L["Background Texture Type"], L["Choose between a Class, Race or Custom Background"], 6, {
 		["CLASS"] = CLASS,
 		["RACE"] = RACE,
@@ -3267,8 +3325,8 @@ function ElvUI_EltreumUI:Configtable()
 	ElvUI_EltreumUI.Options.args.skins.args.character.args.panel.args.armory.args.characterpanelnamefontsize = ACH:Range(L["Character Name Font Size"], nil, 18, { min = 8, max = 40, step = 1 }, "full", function() return E.db.ElvUI_EltreumUI.skins.armorynamefontsize end, function(_, value) E.db.ElvUI_EltreumUI.skins.armorynamefontsize = value ElvUI_EltreumUI:ExpandedCharacterStats() PaperDollFrame_SetLevel() end)
 	ElvUI_EltreumUI.Options.args.skins.args.character.args.panel.args.model = ACH:Group(L["Model"], nil, 1, "tab")
 	ElvUI_EltreumUI.Options.args.skins.args.character.args.panel.args.model.args.description1 = ACH:Description(L["Character Model"], 1, nil, 'Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\EltreumHeader', nil, 3240, 1, "full")
-	ElvUI_EltreumUI.Options.args.skins.args.character.args.panel.args.model.args.modelcamzoom = ACH:Range(L["Character Model Zoom"], nil, 2, { min = -2, max = 2, step = 0.01 }, "full", function() if E.Retail then return E.db.ElvUI_EltreumUI.skins.charactermodelcam.zoomretail else return E.db.ElvUI_EltreumUI.skins.charactermodelcam.zoomclassic end end, function(_, value) if E.Retail then E.db.ElvUI_EltreumUI.skins.charactermodelcam.zoomretail = value else E.db.ElvUI_EltreumUI.skins.charactermodelcam.zoomclassic = value end
-		if E.Retail then
+	ElvUI_EltreumUI.Options.args.skins.args.character.args.panel.args.model.args.modelcamzoom = ACH:Range(L["Character Model Zoom"], nil, 2, { min = -2, max = 2, step = 0.01 }, "full", function() if E.Retail or E.Cata then return E.db.ElvUI_EltreumUI.skins.charactermodelcam.zoomretail else return E.db.ElvUI_EltreumUI.skins.charactermodelcam.zoomclassic end end, function(_, value) if E.Retail or E.Cata then E.db.ElvUI_EltreumUI.skins.charactermodelcam.zoomretail = value else E.db.ElvUI_EltreumUI.skins.charactermodelcam.zoomclassic = value end
+		if E.Retail or E.Cata then
 			if _G.CharacterModelScene then
 				local actor = _G.CharacterModelScene:GetPlayerActor()
 				if actor then
@@ -3283,19 +3341,19 @@ function ElvUI_EltreumUI:Configtable()
 		end
 	end, function() return not E.db.ElvUI_EltreumUI.skins.expandarmorybg end)
 	ElvUI_EltreumUI.Options.args.skins.args.character.args.panel.args.model.args.modelcamx = ACH:Range(L["Character Model X Offset"], nil, 3, { min = -2, max = 2, step = 0.01 }, "full", function()
-		if E.Retail then
+		if E.Retail or E.Cata then
 			return E.db.ElvUI_EltreumUI.skins.charactermodelcam.xretail
 		else
 			return E.db.ElvUI_EltreumUI.skins.charactermodelcam.xclassic
 		end
 	end, function(_, value)
-		if E.Retail then
+		if E.Retail or E.Cata then
 			E.db.ElvUI_EltreumUI.skins.charactermodelcam.xretail = value
 		else
 			E.db.ElvUI_EltreumUI.skins.charactermodelcam.xclassic = value
 		end
 
-		if E.Retail then
+		if E.Retail or E.Cata then
 			if _G.CharacterModelScene then
 				local actor = _G.CharacterModelScene:GetPlayerActor()
 				if actor then
@@ -3310,19 +3368,19 @@ function ElvUI_EltreumUI:Configtable()
 		end
 	end, function() return not E.db.ElvUI_EltreumUI.skins.expandarmorybg end)
 	ElvUI_EltreumUI.Options.args.skins.args.character.args.panel.args.model.args.modelcamy = ACH:Range(L["Character Model Y Offset"], nil, 3, { min = -2, max = 2, step = 0.01 }, "full", function()
-		if E.Retail then
+		if E.Retail or E.Cata then
 			return E.db.ElvUI_EltreumUI.skins.charactermodelcam.yretail
 		else
 			return E.db.ElvUI_EltreumUI.skins.charactermodelcam.yclassic
 		end
 	end, function(_, value)
-		if E.Retail then
+		if E.Retail or E.Cata then
 			E.db.ElvUI_EltreumUI.skins.charactermodelcam.yretail = value
 		else
 			E.db.ElvUI_EltreumUI.skins.charactermodelcam.yclassic = value
 		end
 
-		if E.Retail then
+		if E.Retail or E.Cata then
 			if _G.CharacterModelScene then
 				local actor = _G.CharacterModelScene:GetPlayerActor()
 				if actor then
@@ -3641,28 +3699,20 @@ The Item Level shown on the Character Panel Skin uses code from Simple Item leve
 	ElvUI_EltreumUI.Options.args.changelog.args.added = ACH:Group(E:TextGradient("Added", 0.50, 0.70, 1, 0.67, 0.95, 1), nil, 3)
 	ElvUI_EltreumUI.Options.args.changelog.args.added.inline = true
 	ElvUI_EltreumUI.Options.args.changelog.args.added.args.description = ACH:Description([[
-Added an option to disable moving the Item Level text in the Character Panel in Cataclysm Classic
-Added Eltruism Datatext 3 to Retail where it will show Armor and Dodge change
+Added support for Season of Discovery Season 4
 ]], 3, "small", nil, nil, nil, nil, "full")
 	ElvUI_EltreumUI.Options.args.changelog.args.updated = ACH:Group(E:TextGradient("Updated", 0.50, 0.70, 1, 0.67, 0.95, 1), nil, 3)
 	ElvUI_EltreumUI.Options.args.changelog.args.updated.inline = true
 	ElvUI_EltreumUI.Options.args.changelog.args.updated.args.description = ACH:Description([[
-Updated Aura borders to split Unitframe Auras and Secure Auras
-Updated the color of the texture for Item Level text in Retail
-Updated party gradient to work when not in a group
-Updated Debuff borders to use Color by Dispel
-Updated Debuff borders to use Elvui Colors
-Updated shadows for some new frames
+Updated Item Level text on character panel to follow some of ElvUI's options
+Updated Background Alpha in the Character Panel Skin to not follow Crest Alpha
+Updated Character Panel Gradient and Class Colors to work separately
 ]], 5, "small", nil, nil, nil, nil, "full")
 	ElvUI_EltreumUI.Options.args.changelog.args.fixed = ACH:Group(E:TextGradient("Fixed", 0.50, 0.70, 1, 0.67, 0.95, 1), nil, 4)
 	ElvUI_EltreumUI.Options.args.changelog.args.fixed.inline = true
 	ElvUI_EltreumUI.Options.args.changelog.args.fixed.args.description = ACH:Description([[
-Fixed an issue with the Objective Tracker mover due to a change in Edit Mode
-Fixed an issue with Macro Text colors in Action Bars due to the ElvUI Skin
-Fixed Professions missing in bags in Cataclysm due to different Ids
-Fixed a possible error with the Character Panel in Remix
-Fixed shadows for detached power
-Fixed the zhTW locale missing
+Fixed a possible error when using /way command with text instead of zone id
+Fixed Combat Icons option missing
 ]], 7, "small", nil, nil, nil, nil, "full")
 
 	--[[