+++ ElvUI_EltreumUI/Modules/Misc/InstantLoot.lua
@@ -68,7 +68,7 @@ WishlistItemFrame:Hide()
 -- yet another fast loot thing
 local EltruismInstantLoot = CreateFrame("Frame", "EltruismInstantLoot")
 EltruismInstantLoot:RegisterEvent("LOOT_READY")
-EltruismInstantLoot:RegisterEvent("LOOT_OPENED")
+--EltruismInstantLoot:RegisterEvent("LOOT_OPENED")
 EltruismInstantLoot:RegisterEvent("LOOT_BIND_CONFIRM")
 EltruismInstantLoot:RegisterEvent("LOOT_SLOT_CHANGED")
 EltruismInstantLoot:RegisterEvent("UI_ERROR_MESSAGE")
@@ -84,7 +84,7 @@ local function InstantLoot(_, event,_, arg2)
 		end
 	elseif event == "UI_ERROR_MESSAGE" and arg2 == ERR_INV_FULL then
 		return
-	elseif event == "LOOT_READY" or event == "LOOT_OPENED" or event == "LOOT_SLOT_CHANGED" then
+	elseif (event == "LOOT_READY" or event == "LOOT_SLOT_CHANGED") then
 		--ElvUI_EltreumUI:Print("Event: "..event)
 		if E.db.ElvUI_EltreumUI.loot.lootwishlistwarning then
 			for i = GetNumLootItems(), 1, -1 do
@@ -119,7 +119,6 @@ local function InstantLoot(_, event,_, arg2)
 							E:Delay(0.5, function()
 								lootsoundthrottle = 0
 							end)
-
 						end
 						E:Delay(5, function() UIFrameFadeOut(WishlistItemFrame, 1, 1, 0) end)
 					end