+++ ElvUI_EltreumUI/Modules/Skins/Addons/OmniCD.lua
@@ -7,11 +7,15 @@ local hooksecurefunc = _G.hooksecurefunc
 function ElvUI_EltreumUI:EltruismOmniCD()
 	if E.db.ElvUI_EltreumUI.skins.omnicd then
 
-		--omnicd name bar mode disables draw layer instead, so we need a check for that since api to check for enabled layers doesnt exist it seems
 		local nameBar = false
 		local currentprofile = OmniCDDB["profileKeys"][E.mynameRealm]
-		if OmniCDDB["profiles"][currentprofile]["Party"]["party"]["extraBars"]["raidBar0"]["nameBar"] then
-			nameBar = true
+		if OmniCDDB["profiles"][currentprofile] and
+			OmniCDDB["profiles"][currentprofile]["Party"] and
+			OmniCDDB["profiles"][currentprofile]["Party"]["party"] and
+			OmniCDDB["profiles"][currentprofile]["Party"]["party"]["extraBars"] and
+			OmniCDDB["profiles"][currentprofile]["Party"]["party"]["extraBars"]["raidBar1"] and
+			OmniCDDB["profiles"][currentprofile]["Party"]["party"]["extraBars"]["raidBar1"]["nameBar"] then
+				nameBar = true
 		end
 
 		--omnicd doesnt do normal ace new addon thing
@@ -48,10 +52,10 @@ function ElvUI_EltreumUI:EltruismOmniCD()
 		--interrupt bars shadows/gradient/position
 		hooksecurefunc(OmniCD.Party,"GetStatusBar",function(_, icon)
 			 if icon and icon.statusBar then
-
 				--shadows
 				if E.db.ElvUI_EltreumUI.skins.shadow.enable then
-					if not icon.statusBar.shadow and not nameBar then
+					--omnicd name bar mode disables the BG now instead, so we need a check for that instead of the profile now
+					if icon.statusBar and (not nameBar) and not icon.statusBar.shadow then
 						icon.statusBar:CreateShadow(E.db.ElvUI_EltreumUI.skins.shadow.length)
 						ElvUI_EltreumUI:ShadowColor(icon.statusBar.shadow)
 					end